Error out on non-linux OSes
